requirements.txt
# Packages to install in the pip/conda environment
# output of 'pip freeze > requirements.txt' the 19/06/2024
absl-py==1.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autokeras==1.1.0
backcall==0.2.0
basemap==1.3.6
basemap-data==1.3.2
beautifulsoup4==4.11.2
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cftime==1.6.2
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dask==2023.6.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
executing==1.2.0
fastjsonschema==2.16.3
flatbuffers==23.3.3
fonttools==4.39.0
fqdn==1.5.1
fsspec==2023.6.0
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.11.0
keras-nlp==0.4.1
keras-tuner==1.3.0
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==15.0.6.1
llvmlite==0.42.0
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
netCDF4==1.5.8
notebook==6.5.3
notebook_shim==0.2.2
numba==0.59.0
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyproj==3.4.1
pyrsistent==0.19.3
pyshp==2.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.10.1
Send2Trash==1.8.0
shap==0.45.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.19.0
tensorflow-rocm==2.11.0.540
tensorflow-text==2.11.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
xarray==2023.2.0
xgboost==1.7.4
zipp==3.15.0
